00title almost perfect, a part of 09ready and 01select implemented

+++ gba/types.h
@@ -48,16 +48,16 @@ struct PlttData
 
 struct OamData
 {
-    /*0x00*/ u32 y:8;
-    /*0x01*/ u32 affineMode:2;  // 0x1, 0x2 -> 0x4
-             u32 objMode:2;     // 0x4, 0x8 -> 0xC
-             u32 mosaic:1;      // 0x10
-             u32 bpp:1;         // 0x20
-             u32 shape:2;       // 0x40, 0x80 -> 0xC0
+    /*0x00*/ u16 y:8;
+    /*0x01*/ u16 affineMode:2;  // 0x1, 0x2 -> 0x4
+             u16 objMode:2;     // 0x4, 0x8 -> 0xC
+             u16 mosaic:1;      // 0x10
+             u16 bpp:1;         // 0x20
+             u16 shape:2;       // 0x40, 0x80 -> 0xC0
 
-    /*0x02*/ u32 x:9;
-             u32 matrixNum:5;   // bits 3/4 are h-flip/v-flip if not in affine mode
-             u32 size:2;        // 0x4000, 0x8000 -> 0xC000
+    /*0x02*/ u16 x:9;
+             u16 matrixNum:5;   // bits 3/4 are h-flip/v-flip if not in affine mode
+             u16 size:2;        // 0x4000, 0x8000 -> 0xC000
 
     /*0x04*/ u16 tileNum:10;    // 0x3FF
              u16 priority:2;    // 0x400, 0x800 -> 0xC00
